Detecting Online Fraud: A Data Science Approach to Safeguarding Financial Transactions
Introduction
Fraud is a widespread issue in the digital landscape, affecting areas such as insurance, online auctions, benefit claims, and telecommunication services. Whenever you participate in online transactions, there is a significant risk of encountering fraudulent activity. Therefore, it is crucial to recognize when an online entity is not adhering to the rules.

Given the significant challenges associated with online transactions, you have been employed as a data scientist by a major credit card-issuing bank. Your primary responsibility is to develop a system that leverages purchasing transaction data to swiftly detect fraudulent activities, ideally in real-time. This will enable the bank to implement protective measures for its customers effectively.

For this assignment, you will be working with datasets containing both fraudulent and legitimate purchasing transactions. Your task is to analyze the problem of fraud detection systems and design and implement a solution using the concepts and principles covered in this module.